+++ circuit_maintenance_parser/parser.py
@@ -12,7 +12,7 @@ import hashlib
 import bs4  # type: ignore
 from bs4.element import ResultSet  # type: ignore
 
-from pydantic import BaseModel
+from pydantic import BaseModel, PrivateAttr
 from icalendar import Calendar  # type: ignore
 
 from circuit_maintenance_parser.errors import ParserError
@@ -34,7 +34,7 @@ class Parser(BaseModel):
     """
 
     # _data_types are used to match the Parser to to each type of DataPart
-    _data_types = ["text/plain", "plain"]
+    _data_types = PrivateAttr(["text/plain", "plain"])
 
     # TODO: move it to where it is used, Cogent parser
     _geolocator = Geolocator()
@@ -42,7 +42,7 @@ class Parser(BaseModel):
     @classmethod
     def get_data_types(cls) -> List[str]:
         """Return the expected data type."""
-        return cls._data_types
+        return cls._data_types.get_default()
 
     @classmethod
     def get_name(cls) -> str:
@@ -92,7 +92,7 @@ class ICal(Parser):
     Reference: https://tools.ietf.org/html/draft-gunter-calext-maintenance-notifications-00
     """
 
-    _data_types = ["text/calendar", "ical", "icalendar"]
+    _data_types = PrivateAttr(["text/calendar", "ical", "icalendar"])
 
     def parser_hook(self, raw: bytes, content_type: str):
         """Execute parsing."""
@@ -164,7 +164,7 @@ class ICal(Parser):
 class Html(Parser):
     """Html parser."""
 
-    _data_types = ["text/html", "html"]
+    _data_types = PrivateAttr(["text/html", "html"])
 
     @staticmethod
     def remove_hex_characters(string):
@@ -201,7 +201,11 @@ class Html(Parser):
 class EmailDateParser(Parser):
     """Parser for Email Date."""
 
-    _data_types = [EMAIL_HEADER_DATE]
+    _data_types = PrivateAttr(
+        [
+            EMAIL_HEADER_DATE,
+        ]
+    )
 
     def parser_hook(self, raw: bytes, content_type: str):
         """Execute parsing."""
@@ -214,7 +218,11 @@ class EmailDateParser(Parser):
 class EmailSubjectParser(Parser):
     """Parse data from subject or email."""
 
-    _data_types = [EMAIL_HEADER_SUBJECT]
+    _data_types = PrivateAttr(
+        [
+            EMAIL_HEADER_SUBJECT,
+        ]
+    )
 
     def parser_hook(self, raw: bytes, content_type: str):
         """Execute parsing."""
@@ -236,7 +244,7 @@ class EmailSubjectParser(Parser):
 class Csv(Parser):
     """Csv parser."""
 
-    _data_types = ["application/csv", "text/csv", "application/octet-stream"]
+    _data_types = PrivateAttr(["application/csv", "text/csv", "application/octet-stream"])
 
     def parser_hook(self, raw: bytes, content_type: str):
         """Execute parsing."""
@@ -255,7 +263,11 @@ class Csv(Parser):
 class Text(Parser):
     """Text parser."""
 
-    _data_types = ["text/plain"]
+    _data_types = PrivateAttr(
+        [
+            "text/plain",
+        ]
+    )
 
     def parser_hook(self, raw: bytes, content_type: str):
         """Execute parsing."""
@@ -278,7 +290,7 @@ class Text(Parser):
 class LLM(Parser):
     """LLM parser."""
 
-    _data_types = ["text/html", "html", "text/plain"]
+    _data_types = PrivateAttr(["text/html", "html", "text/plain"])
 
     _llm_question = """Please, could you extract a JSON form without any other comment,
     with the following JSON schema (timestamps in EPOCH and taking into account the GMT offset):