+++ tests/unit/test_e2e.py
@@ -19,11 +19,14 @@ from circuit_maintenance_parser.provider import (
     BSO,
     Cogent,
     Colt,
+    CrownCastle,
     EUNetworks,
+    Google,
     GTT,
     HGC,
     Lumen,
     Megaport,
+    Netflix,
     NTT,
     Momentum,
     PacketFabric,
@@ -100,6 +103,15 @@ GENERIC_ICAL_RESULT_PATH = Path(dir_path, "data", "ical", "ical1_result.json")
             ],
             [
                 Path(dir_path, "data", "aws", "aws2_result.json"),
+            ],
+        ),
+        (
+            AWS,
+            [
+                ("email", Path(dir_path, "data", "aws", "aws3.eml")),
+            ],
+            [
+                Path(dir_path, "data", "aws", "aws3_result.json"),
             ],
         ),
         # BSO
@@ -249,6 +261,21 @@ GENERIC_ICAL_RESULT_PATH = Path(dir_path, "data", "ical", "ical1_result.json")
                 Path(dir_path, "data", "colt", "colt7_result.json"),
             ],
         ),
+        # Crown Castle
+        (
+            CrownCastle,
+            [
+                ("email", Path(dir_path, "data", "crowncastle", "crowncastle1.eml")),
+            ],
+            [Path(dir_path, "data", "crowncastle", "crowncastle1_result.json")],
+        ),
+        (
+            CrownCastle,
+            [
+                ("email", Path(dir_path, "data", "crowncastle", "crowncastle7.eml")),
+            ],
+            [Path(dir_path, "data", "crowncastle", "crowncastle7_result.json")],
+        ),
         # Equinix
         (
             Equinix,
@@ -269,6 +296,26 @@ GENERIC_ICAL_RESULT_PATH = Path(dir_path, "data", "ical", "ical1_result.json")
             Equinix,
             [("email", Path(dir_path, "data", "equinix", "equinix5.eml"))],
             [Path(dir_path, "data", "equinix", "equinix5_result_combined.json")],
+        ),
+        (
+            Equinix,
+            [("email", Path(dir_path, "data", "equinix", "equinix6.eml"))],
+            [Path(dir_path, "data", "equinix", "equinix6_result_combined.json")],
+        ),
+        (
+            Equinix,
+            [("email", Path(dir_path, "data", "equinix", "equinix7.eml"))],
+            [Path(dir_path, "data", "equinix", "equinix7_result_combined.json")],
+        ),
+        (
+            Equinix,
+            [("email", Path(dir_path, "data", "equinix", "equinix8.eml"))],
+            [Path(dir_path, "data", "equinix", "equinix8_result_combined.json")],
+        ),
+        (
+            Equinix,
+            [("email", Path(dir_path, "data", "equinix", "equinix9.eml"))],
+            [Path(dir_path, "data", "equinix", "equinix9_result_combined.json")],
         ),
         # EUNetworks
         (
@@ -372,6 +419,16 @@ GENERIC_ICAL_RESULT_PATH = Path(dir_path, "data", "ical", "ical1_result.json")
             [
                 Path(dir_path, "data", "hgc", "hgc1_result.json"),
                 Path(dir_path, "data", "hgc", "hgc2_result.json"),
+            ],
+        ),
+        # Google
+        (
+            Google,
+            [
+                ("email", Path(dir_path, "data", "google", "google2.eml")),
+            ],
+            [
+                Path(dir_path, "data", "google", "google2_result.json"),
             ],
         ),
         # Lumen
@@ -501,6 +558,25 @@ GENERIC_ICAL_RESULT_PATH = Path(dir_path, "data", "ical", "ical1_result.json")
             ],
             [
                 Path(dir_path, "data", "momentum", "momentum1_result.json"),
+            ],
+        ),
+        # Netflix
+        (
+            Netflix,
+            [
+                ("email", Path(dir_path, "data", "netflix", "netflix1.eml")),
+            ],
+            [
+                Path(dir_path, "data", "netflix", "netflix1_result.json"),
+            ],
+        ),
+        (
+            Netflix,
+            [
+                ("email", Path(dir_path, "data", "netflix", "netflix2.eml")),
+            ],
+            [
+                Path(dir_path, "data", "netflix", "netflix2_result.json"),
             ],
         ),
         # NTT
@@ -838,7 +914,6 @@ def test_provider_get_maintenances(
     extended_data = provider_class.get_extended_data()
     default_maintenance_data = {"uid": "0", "sequence": 1, "summary": ""}
     extended_data.update(default_maintenance_data)
-
     data = None
     for data_type, data_file in test_data_files:
         with open(data_file, "rb") as file_obj:
@@ -891,8 +966,7 @@ def test_provider_get_maintenances(
             """\
 Failed creating Maintenance notification for GenericProvider.
 Details:
-- Processor SimpleProcessor from GenericProvider failed due to: 1 validation error for Maintenance\naccount\n  field required (type=value_error.missing)
-""",
+- Processor SimpleProcessor from GenericProvider failed due to: 1 validation error for Maintenance\naccount\n  Field required""",
         ),
         (
             GenericProvider,
@@ -904,8 +978,7 @@ Failed creating Maintenance notification for GenericProvider.
 Details:
 - Processor SimpleProcessor from GenericProvider failed due to: 1 validation error for Maintenance
 maintenance_id
-  field required (type=value_error.missing)
-""",
+  Field required""",
         ),
         (
             GenericProvider,
@@ -945,13 +1018,11 @@ Details:
             "ical",
             Path(dir_path, "data", "ical", "ical_no_account"),
             ProviderError,
-            """\
-Failed creating Maintenance notification for Telstra.
-Details:
-- Processor SimpleProcessor from Telstra failed due to: 1 validation error for Maintenance\naccount\n  field required (type=value_error.missing)
-- Processor CombinedProcessor from Telstra failed due to: None of the supported parsers for processor CombinedProcessor (EmailDateParser, HtmlParserTelstra2) was matching any of the provided data types (ical).
-- Processor CombinedProcessor from Telstra failed due to: None of the supported parsers for processor CombinedProcessor (EmailDateParser, HtmlParserTelstra1) was matching any of the provided data types (ical).
-""",
+            [
+                "- Processor SimpleProcessor from Telstra failed due to: 1 validation error for Maintenance\naccount\n  Field required",
+                "- Processor CombinedProcessor from Telstra failed due to: None of the supported parsers for processor CombinedProcessor (EmailDateParser, HtmlParserTelstra2) was matching any of the provided data types (ical).",
+                "- Processor CombinedProcessor from Telstra failed due to: None of the supported parsers for processor CombinedProcessor (EmailDateParser, HtmlParserTelstra1) was matching any of the provided data types (ical).",
+            ],
         ),
         # Zayo
         (
@@ -964,8 +1035,7 @@ Failed creating Maintenance notification for Zayo.
 Details:
 - Processor CombinedProcessor from Zayo failed due to: 1 validation error for Maintenance
 maintenance_id
-  String is empty or 'None' (type=value_error)
-""",
+  Value error, String is empty or 'None'""",
         ),
         (
             Zayo,
@@ -998,5 +1068,10 @@ def test_errored_provider_process(provider_class, data_type, data_file, exceptio
     with pytest.raises(exception) as exc:
         provider_class().get_maintenances(data)
 
-    assert len(exc.value.related_exceptions) == len(provider_class._processors)  # pylint: disable=protected-access
-    assert str(exc.value) == error_message
+    assert len(exc.value.related_exceptions) == len(
+        provider_class.get_default_processors()
+    )  # pylint: disable=protected-access
+    if isinstance(error_message, str):
+        error_message = [error_message]
+    for item in error_message:
+        assert item in str(exc.value)