+++ tests/unit/test_parsers.py
@@ -12,12 +12,15 @@ from circuit_maintenance_parser.parsers.aws import SubjectParserAWS1, TextParser
 from circuit_maintenance_parser.parsers.bso import HtmlParserBSO1
 from circuit_maintenance_parser.parsers.cogent import HtmlParserCogent1
 from circuit_maintenance_parser.parsers.colt import CsvParserColt1, SubjectParserColt1, SubjectParserColt2
+from circuit_maintenance_parser.parsers.crowncastle import HtmlParserCrownCastle1
 from circuit_maintenance_parser.parsers.equinix import HtmlParserEquinix, SubjectParserEquinix
+from circuit_maintenance_parser.parsers.google import HtmlParserGoogle1
 from circuit_maintenance_parser.parsers.gtt import HtmlParserGTT1
 from circuit_maintenance_parser.parsers.hgc import HtmlParserHGC1, HtmlParserHGC2
 from circuit_maintenance_parser.parsers.lumen import HtmlParserLumen1
 from circuit_maintenance_parser.parsers.megaport import HtmlParserMegaport1
 from circuit_maintenance_parser.parsers.momentum import HtmlParserMomentum1
+from circuit_maintenance_parser.parsers.netflix import TextParserNetflix1
 from circuit_maintenance_parser.parsers.seaborn import (
     HtmlParserSeaborn1,
     HtmlParserSeaborn2,
@@ -32,6 +35,20 @@ from circuit_maintenance_parser.parsers.zayo import SubjectParserZayo1, HtmlPars
 
 
 dir_path = os.path.dirname(os.path.realpath(__file__))
+
+
+# TODO: As a future breaking change, the output could be fully serializable by default
+# Currently, the object contains internal objects (e.g., CircuitImpact) that can't be
+# converted into JSON without this encoder.
+class NestedEncoder(json.JSONEncoder):
+    """Helper Class to encode to JSON recursively calling custom methods."""
+
+    def default(self, o):
+        """Expected method to serialize to JSON."""
+        if hasattr(o, "to_json"):
+            return o.to_json()
+
+        return json.JSONEncoder.default(self, o)
 
 
 @pytest.mark.parametrize(
@@ -104,6 +121,11 @@ dir_path = os.path.dirname(os.path.realpath(__file__))
             SubjectParserAWS1,
             Path(dir_path, "data", "aws", "aws2.eml"),
             Path(dir_path, "data", "aws", "aws2_subject_parser_result.json"),
+        ),
+        (
+            TextParserAWS1,
+            Path(dir_path, "data", "aws", "aws3.eml"),
+            Path(dir_path, "data", "aws", "aws3_text_parser_result.json"),
         ),
         # BSO
         (
@@ -213,6 +235,37 @@ dir_path = os.path.dirname(os.path.realpath(__file__))
             Path(dir_path, "data", "colt", "colt7.eml"),
             Path(dir_path, "data", "colt", "colt7_subject_parser_1_result.json"),
         ),
+        # Crown Castle Fiber
+        (
+            HtmlParserCrownCastle1,
+            Path(dir_path, "data", "crowncastle", "crowncastle2.html"),
+            Path(dir_path, "data", "crowncastle", "crowncastle2_parser_result.json"),
+        ),
+        (
+            HtmlParserCrownCastle1,
+            Path(dir_path, "data", "crowncastle", "crowncastle3.html"),
+            Path(dir_path, "data", "crowncastle", "crowncastle3_parser_result.json"),
+        ),
+        (
+            HtmlParserCrownCastle1,
+            Path(dir_path, "data", "crowncastle", "crowncastle4.html"),
+            Path(dir_path, "data", "crowncastle", "crowncastle4_parser_result.json"),
+        ),
+        (
+            HtmlParserCrownCastle1,
+            Path(dir_path, "data", "crowncastle", "crowncastle5.html"),
+            Path(dir_path, "data", "crowncastle", "crowncastle5_parser_result.json"),
+        ),
+        (
+            HtmlParserCrownCastle1,
+            Path(dir_path, "data", "crowncastle", "crowncastle6.html"),
+            Path(dir_path, "data", "crowncastle", "crowncastle6_parser_result.json"),
+        ),
+        (
+            HtmlParserCrownCastle1,
+            Path(dir_path, "data", "crowncastle", "crowncastle8.html"),
+            Path(dir_path, "data", "crowncastle", "crowncastle8_parser_result.json"),
+        ),
         # Equinix
         (
             HtmlParserEquinix,
@@ -238,6 +291,32 @@ dir_path = os.path.dirname(os.path.realpath(__file__))
             HtmlParserEquinix,
             Path(dir_path, "data", "equinix", "equinix5.eml"),
             Path(dir_path, "data", "equinix", "equinix5_result.json"),
+        ),
+        (
+            HtmlParserEquinix,
+            Path(dir_path, "data", "equinix", "equinix6.eml"),
+            Path(dir_path, "data", "equinix", "equinix6_result.json"),
+        ),
+        (
+            HtmlParserEquinix,
+            Path(dir_path, "data", "equinix", "equinix7.eml"),
+            Path(dir_path, "data", "equinix", "equinix7_result.json"),
+        ),
+        (
+            HtmlParserEquinix,
+            Path(dir_path, "data", "equinix", "equinix8.eml"),
+            Path(dir_path, "data", "equinix", "equinix8_result.json"),
+        ),
+        (
+            HtmlParserEquinix,
+            Path(dir_path, "data", "equinix", "equinix9.eml"),
+            Path(dir_path, "data", "equinix", "equinix9_result.json"),
+        ),
+        # Google
+        (
+            HtmlParserGoogle1,
+            Path(dir_path, "data", "google", "google1.html"),
+            Path(dir_path, "data", "google", "google1_parser_result.json"),
         ),
         # GTT
         (
@@ -343,6 +422,17 @@ dir_path = os.path.dirname(os.path.realpath(__file__))
             HtmlParserMomentum1,
             Path(dir_path, "data", "momentum", "momentum1.eml"),
             Path(dir_path, "data", "momentum", "momentum1_html_parser_result.json"),
+        ),
+        # Netflix
+        (
+            TextParserNetflix1,
+            Path(dir_path, "data", "netflix", "netflix1.eml"),
+            Path(dir_path, "data", "netflix", "netflix1_text_parser_result.json"),
+        ),
+        (
+            TextParserNetflix1,
+            Path(dir_path, "data", "netflix", "netflix2.eml"),
+            Path(dir_path, "data", "netflix", "netflix2_text_parser_result.json"),
         ),
         # NTT
         (
@@ -551,6 +641,8 @@ def test_parsers(parser_class, raw_file, results_file):
 
     parsed_notifications = parser_class().parse(raw_data, parser_class.get_data_types()[0])
 
+    parsed_notifications = json.loads(json.dumps(parsed_notifications, cls=NestedEncoder))
+
     with open(results_file, encoding="utf-8") as res_file:
         expected_result = json.load(res_file)
 
@@ -564,4 +656,4 @@ def test_parsers(parser_class, raw_file, results_file):
 def test_parser_no_data(parser_class):
     """Test parser with no data."""
     with pytest.raises(ParserError):
-        parser_class().parse(b"", parser_class._data_types[0])  # pylint: disable=protected-access
+        parser_class().parse(b"", parser_class.get_data_types()[0])  # pylint: disable=protected-access